feat(notewave): rename ClassDashboard, add topic page for student and teacher
