+++ src/app/[locale]/class-based/[id]/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useClassBased } from '@/contexts/class-based';
-import ClassDashboard from './dashboard/ClassDashboard';
+import StudentClassDashboard from './dashboard/StudentClassDashboard';
 
 export default function Page() {
     const { classId } = useClassBased();
@@ -12,7 +12,7 @@ export default function Page() {
 
     return (
         <div className="flex flex-col h-full mt-4">
-            <ClassDashboard />
+            <StudentClassDashboard />
         </div>
     );
 }
